Refactor Login page (#84)
