+++ src/pages/SignUp/index.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function SignUp() {
+  return <div>Hello, World!</div>;
+}
+
+export default SignUp;
